+++ src/Trading.Bot.API/Mediator/RsiEmaHandler.cs
@@ -8,9 +8,9 @@ public class RsiEmaRequestHandler : IRequestHandler<RsiEmaRequest, IResult>
 
         var rsiLimit = request.RsiLimit ?? 50;
 
-        var maxSpread = request.MaxSpread ?? 0.0003;
+        var maxSpread = request.MaxSpread ?? 0.0003m;
 
-        var minGain = request.MinGain ?? 0.0006;
+        var minGain = request.MinGain ?? 0.0006m;
 
         var riskReward = request.RiskReward ?? 1;
 
@@ -46,8 +46,8 @@ public record RsiEmaRequest : IHttpRequest
     public int RsiWindow { get; set; }
     public int EmaWindow { get; set; }
     public double? RsiLimit { get; set; }
-    public double? MaxSpread { get; set; }
-    public double? MinGain { get; set; }
-    public double? RiskReward { get; set; }
+    public decimal? MaxSpread { get; set; }
+    public decimal? MinGain { get; set; }
+    public decimal? RiskReward { get; set; }
     public int? TradeRisk { get; set; }
 }
