+++ src/Trading.Bot.API/Models/TradeResult.cs
@@ -4,13 +4,13 @@ public class TradeResult
 {
     public bool Running { get; set; }
     public int StartIndex { get; set; }
-    public double StartPrice { get; set; }
-    public double TriggerPrice { get; set; }
+    public decimal StartPrice { get; set; }
+    public decimal TriggerPrice { get; set; }
     public Signal Signal { get; set; }
-    public double TakeProfit { get; set; }
-    public double StopLoss { get; set; }
-    public double UnrealisedPL { get; set; }
+    public decimal TakeProfit { get; set; }
+    public decimal StopLoss { get; set; }
+    public decimal UnrealisedPL { get; set; }
     public DateTime StartTime { get; set; }
     public DateTime EndTime { get; set; }
-    public double Result { get; set; }
+    public decimal Result { get; set; }
 }
