+++ src/Trading.Bot/Models/ApiResponses/AccountResponse.cs
@@ -10,30 +10,30 @@ public class AccountResponse
     public string Currency { get; set; }
     public int CreatedByUserID { get; set; }
     public string Alias { get; set; }
-    public double MarginRate { get; set; }
+    public decimal MarginRate { get; set; }
     public string LastTransactionID { get; set; }
-    public double Balance { get; set; }
+    public decimal Balance { get; set; }
     public int OpenTradeCount { get; set; }
     public int OpenPositionCount { get; set; }
     public int PendingOrderCount { get; set; }
-    public double Pl { get; set; }
-    public double ResettablePL { get; set; }
+    public decimal Pl { get; set; }
+    public decimal ResettablePL { get; set; }
     public string ResettablePLTime { get; set; }
-    public double Financing { get; set; }
-    public double Commission { get; set; }
-    public double DividendAdjustment { get; set; }
-    public double GuaranteedExecutionFees { get; set; }
-    public double UnrealizedPL { get; set; }
-    public double NAV { get; set; }
-    public double MarginUsed { get; set; }
-    public double MarginAvailable { get; set; }
-    public double PositionValue { get; set; }
-    public double MarginCloseoutUnrealizedPL { get; set; }
-    public double MarginCloseoutNAV { get; set; }
-    public double MarginCloseoutMarginUsed { get; set; }
-    public double MarginCloseoutPositionValue { get; set; }
-    public double MarginCloseoutPercent { get; set; }
-    public double WithdrawalLimit { get; set; }
-    public double MarginCallMarginUsed { get; set; }
-    public double MarginCallPercent { get; set; }
+    public decimal Financing { get; set; }
+    public decimal Commission { get; set; }
+    public decimal DividendAdjustment { get; set; }
+    public decimal GuaranteedExecutionFees { get; set; }
+    public decimal UnrealizedPL { get; set; }
+    public decimal NAV { get; set; }
+    public decimal MarginUsed { get; set; }
+    public decimal MarginAvailable { get; set; }
+    public decimal PositionValue { get; set; }
+    public decimal MarginCloseoutUnrealizedPL { get; set; }
+    public decimal MarginCloseoutNAV { get; set; }
+    public decimal MarginCloseoutMarginUsed { get; set; }
+    public decimal MarginCloseoutPositionValue { get; set; }
+    public decimal MarginCloseoutPercent { get; set; }
+    public decimal WithdrawalLimit { get; set; }
+    public decimal MarginCallMarginUsed { get; set; }
+    public decimal MarginCallPercent { get; set; }
 }
